+++ service/src/main/java/org/ehrbase/ehr/knowledge/I_KnowledgeCache.java
@@ -21,6 +21,13 @@
  */
 package org.ehrbase.ehr.knowledge;
 
+import java.io.IOException;
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+
 import org.ehrbase.api.exception.InternalServerException;
 import org.ehrbase.api.exception.InvalidApiParameterException;
 import org.ehrbase.api.exception.StateConflictException;
@@ -73,6 +80,24 @@ public interface I_KnowledgeCache {
      */
     Optional<OPERATIONALTEMPLATE> retrieveOperationalTemplate(UUID uuid);
 
+    /**
+     * Checks whether the template contains all nodeIds
+     *
+     * @param templateId
+     * @param nodeIds
+     * @return
+     */
+    boolean containsNodeIds(String templateId, Collection<String> nodeIds);
+
+    /**
+     * Deletes a given operational template physically from cache and from template storage and from cache. Should only
+     * be executed if the template is no longer referenced by any Composition. Make sure you check for references before
+     * deleting a template otherwise this causes inconsistencies and no longer deliverable Composition entries.
+     *
+     * @param template - The template instance to delete
+     * @return - Template has been deleted
+     */
+    boolean deleteOperationalTemplate(OPERATIONALTEMPLATE template);
 
     JsonPathQueryResult resolveForTemplate(String templateId, String jsonQueryExpression);
 }