+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add admin API endpoint stubs (see: https://github.com/ehrbase/ehrbase/pull/280)
 - Add support for FeederAudit in Locatable. Refactored Composition Serializer for DB encoding (see https://github.com/ehrbase/ehrbase/tree/feature/311_feeder_audit, https://github.com/ehrbase/openEHR_SDK/tree/feature/311_feeder_audit)
 - Change the strategy to resolve CONTAINS in AQL (https://github.com/ehrbase/ehrbase/pull/276)
+- Add admin template API functionality (see: https://github.com/ehrbase/ehrbase/pull/301)
 - Persist caches to java.io.tmpdir  (see: https://github.com/ehrbase/ehrbase/pull/308)
 - Precalculate containment tree from OPT template (see https://github.com/ehrbase/ehrbase/pull/312)